--- src/ibl_to_nwb/brainwide_map/__init__.py
@@ -1,1 +1,0 @@
-from .brainwidemapconverter import BrainwideMapConverter
