--- src/ibl_to_nwb/brainwide_map/datainterfaces/__init__.py
@@ -1,1 +1,0 @@
-from .brainwidemaptrialsinterface import BrainwideMapTrialsInterface
