+++ src/ibl_to_nwb/datainterfaces/brainwidemaptrialsinterface.py
