+++ src/ibl_to_nwb/environments/initial_conversion_environment.txt
