+++ src/ibl_to_nwb/metadata/brainwide_map_general.yml
