+++ src/ibl_to_nwb/metadata/trials.yml
