--- src/ibl_to_nwb/requirements.txt
@@ -1,8 +1,0 @@
-neuroconv
-spikeinterface>=0.97.0  # Actually need to use the streaming IBL fix for maximum efficiency
-ONE-api>=1.16.3
-ibllib>=2.21.0
-ndx-pose>=0.1.1
-ndx-ibl==0.1.0
-probeinterface
-iblatlas
