+++ src/ibl_to_nwb/scripts/convert_brainwide_map.py
